MovieCollection/requirements.txt
requests
environ
requests
djangorestframework
djangorestframework-simplejwt